+++ src/main/webapp/resources/js/pages/projects/settings/index.js
@@ -1,8 +1,14 @@
-import { Redirect, Router } from "@reach/router";
-import { Col, Layout, Row, Skeleton } from "antd";
-import React, { Suspense } from "react";
+import { Col, Layout, Row, Spin } from "antd";
+import React, { Suspense, useState } from "react";
 import { render } from "react-dom";
 import { Provider } from "react-redux";
+import {
+  BrowserRouter,
+  Outlet,
+  Route,
+  Routes,
+  useParams,
+} from "react-router-dom";
 import { useGetProjectDetailsQuery } from "../../../apis/projects/project";
 import { MetadataRolesProvider } from "../../../contexts/metadata-roles-context";
 import { ProjectRolesProvider } from "../../../contexts/project-roles-context";
@@ -11,6 +17,8 @@ import { SPACE_SM } from "../../../styles/spacing";
 import { setBaseUrl } from "../../../utilities/url-utilities";
 import SettingsNav from "./components/SettingsNav";
 import store from "./store";
+import { getProjectRoles } from "../../../apis/projects/projects";
+import { useMetadataRoles } from "../../../contexts/metadata-roles-context";
 
 const ProjectDetails = React.lazy(() => import("./components/ProjectDetails"));
 const ProjectProcessing = React.lazy(() =>
@@ -25,22 +33,24 @@ const MetadataLayout = React.lazy(() =>
   import("./components/metadata/MetadataLayout")
 );
 
-const MetadataFieldsListManager = React.lazy(() =>
-  import("./components/metadata/MetadataFieldsListManager")
+// const MetadataFieldsListManager = React.lazy(() =>
+//   import("./components/metadata/MetadataFieldsListManager")
+// );
+//
+// const MetadataFieldsListMember = React.lazy(() =>
+//   import("./components/metadata/MetadataFieldsListMember")
+// );
+
+const MetadataFields = React.lazy(() =>
+  import("./components/metadata/MetadataFields")
 );
 
-const MetadataFieldsListMember = React.lazy(() =>
-  import("./components/metadata/MetadataFieldsListMember")
-);
-
-const MetadataTemplateManager = React.lazy(() =>
-  import("./components/metadata/MetadataTemplateManager")
-);
-const MetadataTemplateMember = React.lazy(() =>
-  import("./components/metadata/MetadataTemplateMember")
-);
 const MetadataTemplates = React.lazy(() =>
   import("./components/metadata/MetadataTemplates")
+);
+
+const MetadataTemplate = React.lazy(() =>
+  import("./components/metadata/MetadataTemplate")
 );
 
 const AssociatedProjects = React.lazy(() =>
@@ -76,9 +86,27 @@ const { Content, Sider } = Layout;
  * @constructor
  */
 const SettingsLayout = () => (
-  <Router>
-    <ProjectSettings path={setBaseUrl("/projects/:projectId/settings/*")} />
-  </Router>
+  <Routes>
+    <Route
+      path={setBaseUrl("/projects/:projectId/settings")}
+      element={<ProjectSettings />}
+    >
+      <Route index element={<ProjectDetails />} />
+      <Route path="processing" element={<ProjectProcessing />} />
+      <Route path="members" element={<ProjectMembers />} />
+      <Route path="groups" element={<ProjectGroups />} />
+      <Route path="metadata" element={<MetadataLayout />}>
+        <Route path="fields" element={<MetadataFields />} />
+        <Route path="templates" element={<MetadataTemplates />} />
+        <Route path="templates/:id" element={<MetadataTemplate />} />
+      </Route>
+      <Route path="associated" element={<AssociatedProjects />} />
+      <Route path="references" element={<ReferenceFiles />} />
+      <Route path="remote" element={<ProjectSynchronizationSettings />} />
+      <Route path="delete" element={<DeleteProject />} />
+      <Route path="*" element={<ProjectDetails />} />
+    </Route>
+  </Routes>
 );
 
 /**
@@ -87,13 +115,22 @@ const SettingsLayout = () => (
  * @returns {JSX.Element}
  * @constructor
  */
-const ProjectSettings = (props) => {
-  const { data: project = {} } = useGetProjectDetailsQuery(props.projectId);
+const ProjectSettings = () => {
+  const { projectId } = useParams();
+
+  const { data: project = {} } = useGetProjectDetailsQuery(projectId, {
+    skip: !projectId,
+  });
+
+  const [basePath] = useState(() =>
+    setBaseUrl(`/projects/${projectId}/settings/`)
+  );
+
   return (
     <Layout>
       <Sider width={200} style={{ backgroundColor: grey1 }}>
         <SettingsNav
-          path={props["*"]}
+          basePath={basePath}
           canManage={project.canManage}
           showRemote={project.canManage && project.remote}
         />
@@ -102,34 +139,10 @@ const ProjectSettings = (props) => {
         <Content style={{ backgroundColor: grey1, paddingLeft: SPACE_SM }}>
           <Row>
             <Col lg={24} xxl={12}>
-              <ProjectRolesProvider>
-                <MetadataRolesProvider>
-                  <Suspense fallback={<Skeleton />}>
-                    <Router>
-                      <ProjectDetails path="/details" />
-                      <ProjectProcessing path="/processing" />
-                      <ProjectMembers path="/members" />
-                      <ProjectGroups path="/groups" />
-                      <MetadataLayout path="/metadata">
-                        <MetadataTemplates path="/templates" />
-                        {project.canManage ? (
-                          <>
-                            <MetadataFieldsListManager path="/fields" />
-                            <MetadataTemplateManager path="/templates/:id" />
-                          </>
-                        ) : (
-                          <>
-                            <MetadataFieldsListMember path="/fields" />
-                            <MetadataTemplateMember path="/templates/:id" />
-                          </>
-                        )}
-                      </MetadataLayout>
-                      <AssociatedProjects path="/associated" />
-                      <ReferenceFiles path="/references" />
-                      <ProjectSynchronizationSettings path="/remote" />
-                      {project.canManage && <DeleteProject path="/delete" />}
-                      <Redirect from="/" to="/details" />
-                    </Router>
+              <ProjectRolesProvider getRolesFn={getProjectRoles}>
+                <MetadataRolesProvider getRolesFn={useMetadataRoles}>
+                  <Suspense fallback={<Spin />}>
+                    <Outlet />
                   </Suspense>
                 </MetadataRolesProvider>
               </ProjectRolesProvider>
@@ -142,8 +155,10 @@ const ProjectSettings = (props) => {
 };
 
 render(
-  <Provider store={store}>
-    <SettingsLayout />
-  </Provider>,
+  <BrowserRouter>
+    <Provider store={store}>
+      <SettingsLayout />
+    </Provider>
+  </BrowserRouter>,
   document.querySelector("#root")
 );