+++ src/test/java/ca/corefacility/bioinformatics/irida/pipeline/results/updater/impl/unit/SISTRSampleUpdaterTest.java
@@ -31,8 +31,8 @@ import java.util.Set;
 import java.util.UUID;
 
 import static org.junit.Assert.assertEquals;
-import static org.mockito.Matchers.any;
-import static org.mockito.Matchers.eq;
+import static org.mockito.ArgumentMatchers.any;
+import static org.mockito.ArgumentMatchers.eq;
 import static org.mockito.Mockito.*;
 
 public class SISTRSampleUpdaterTest {