+++ src/test/java/ca/corefacility/bioinformatics/irida/processing/impl/unit/FastqcFileProcessorTest.java
@@ -1,7 +1,7 @@
 package ca.corefacility.bioinformatics.irida.processing.impl.unit;
 
 import static org.junit.Assert.*;
-import static org.mockito.Matchers.any;
+import static org.mockito.ArgumentMatchers.any;
 import static org.mockito.Mockito.*;
 
 import java.io.IOException;