+++ src/test/java/ca/corefacility/bioinformatics/irida/ria/unit/web/SequencingRunSequenceFilesControllerTest.java
@@ -24,7 +24,7 @@ import java.util.Map;
 
 import static org.junit.Assert.*;
 import static org.mockito.Mockito.*;
-import static org.springframework.hateoas.server.mvc.ControllerLinkBuilder.linkTo;
+import static org.springframework.hateoas.server.mvc.WebMvcLinkBuilder.linkTo;
 
 /**
  * Tests for {@link RESTSequencingRunSequenceFilesController}.